Simplify query for loading all open CFD IDs
Since we now periodically move CFDs that are failed and closed to
separate tables, we don't really need to run a relatively expensive
query every time we want to get open CFDs.

This does mean that sometimes we may return CFDs that are no longer
relevant, but we are willing to accept these transient inconsistencies
in the name of performance.