WIP add steps for initial tour
Manerged to auto-novigate when stepping through the application, but it does not work properly.

Problems: When navigating to a different route info.js cannot handle correct rendering of the steps because it appears that the position is decided based on the initial render upon loading.
Currently not sure how to overcome that.